210712_커뮤니티 관련 기능 추가

+++ app_main/src/main/java/app/jw/mapable/gm/community/ItemCommunityMain.kt
@@ -7,8 +7,8 @@ data class ItemCommunityMain (
     val posttime : String,
     val imageLink : String,
 
-    val like : Int,
-    val dislike : Int,
-    val star : Int,
+    val like : Long,
+    val dislike : Long,
+    val star : Long,
 
 )
